refactor : layout 디렉토리 변경
