+++ client/src/views/AggregateCategoryPage/index.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import MainLayout from '@/components/common/layouts/MainLayout';
+import MainLayout from '@/views/layouts/MainLayout';
 import AggregateCategory from '@/container/AggregateCategory';
 import SelectMonth from '@/container/SelectMonth';
