+++ client/src/views/layouts/MainLayout/index.tsx
@@ -1,5 +1,5 @@
 import React, { useCallback, useEffect, useState } from 'react';
-import Header from '@/components/common/layouts/Header';
+import Header from '@/views/layouts/Header';
 import GitHubSVG from '@/assets/svg/GitHub.svg';
 import { useHistory } from 'react-router-dom';
 import { useSelector, useDispatch } from 'react-redux';
@@ -8,7 +8,7 @@ import authorizationAPI from '@/libs/api/Authorization';
 import { login } from '@/modules/authorization/actions';
 import * as S from './styles';
 import { MainLayoutPropsType } from './types';
-import LoadingSpinner from '../../LoadingSpinner';
+import LoadingSpinner from '../../../components/common/LoadingSpinner';
 
 const MainLayout = ({ children }: MainLayoutPropsType): JSX.Element => {
   const [loading, setLoading] = useState(true);
